EnemyBehavior update
apply EnemyBehavior for Spider and Flower, eliminate Enemy Health system in Game Manager, use PlayerPrefs to store enemy's data, fix Flower animation

+++ Assets/EnemyHealthBar.cs
@@ -13,10 +13,9 @@ public class EnemyHealthBar : MonoBehaviour
     void Update()
     {
         slider.transform.position = Camera.main.WorldToScreenPoint(transform.parent.position + offSet);
-        
     }
 
-    public void SetHealth(int health, int maxHealth)
+    public void SetHealth(float health, float maxHealth)
     {
         slider.gameObject.SetActive(health < maxHealth && health>0);
         slider.maxValue = maxHealth;
